// src/services/types.ts
export interface data{
    total_pages: number,
    results:[results]
}


export interface results{
    backdrop_path: string
    genre_ids: number[]
    id: number
    original_language: string
    original_title: string
    overview: string
    popularity: number
    poster_path: string
    release_date: string
    title: string
    video: false
    vote_average: number
    vote_count: number

}